fix - created different entry points for prod/dev to remove react-hot-loader dependency
