+++ packages/blockchain-wallet-v4-frontend/src/index.dev.js
@@ -1,0 +1,26 @@
+import './assets/sass/global.scss'
+import React from 'react'
+import ReactDOM from 'react-dom'
+import { AppContainer } from 'react-hot-loader'
+import App from 'scenes/app.js'
+import configureStore from 'store'
+import configureLocales from 'services/LocalesService'
+
+const { store, history } = configureStore()
+
+const { messages } = configureLocales(store)
+
+const render = Component => {
+  ReactDOM.render(
+    <AppContainer key={Math.random()}>
+      <Component store={store} history={history} messages={messages} />
+    </AppContainer>,
+    document.getElementById('app')
+  )
+}
+
+render(App)
+
+if (module.hot) {
+  module.hot.accept('scenes/app.js', () => render(require('scenes/app.js').default))
+}
