+++ packages/blockchain-wallet-v4-frontend/src/store/index.js
@@ -1,4 +1,4 @@
-let configureStore = (process.env.NODE_ENV === 'production')
+const configureStore = (process.env.NODE_ENV === 'production')
   ? require('./configureStore.prod')
   : require('./configureStore.dev')
